use sentinel mechanic to run upgrade scripts
When the package wazo-upgrade is installed from the first time, then all
sentinel files are created and no script is executed. We don't care of
the wazo version installed, we only care of the existence of the
sentinel file. This resolve the following problem:

Given doing a wazo-upgrade
When the wazo-upgrade post-start script fail
Then I cannot re-run wazo-upgrade
Expected: I can fix why the script has failed and re-run wazo-upgrade
Cause: the script check the wazo version and wazo has been updated

This also simplify the dev version of wazo, because all dev VM are
already in the next version, so there are no script executed

+++ post-stop.d/50-remove-nginx-symlinks.sh
@@ -9,4 +9,10 @@ set -o pipefail  # fail if command before pipe fails
 # This symlink is left over, after removing wazo-admin-ui and prevents nginx
 # from loading correctly.
 
+SENTINEL="/var/lib/wazo-upgrade/remove-nginx-symlinks"
+
+[ -e "$SENTINEL" ] && exit 0
+
 rm -f /etc/nginx/locations/https-enabled/wazo-admin-ui
+
+touch $SENTINEL
